# server/portal/internal/docs/swagger.yaml
basePath: /fe-v1
definitions:
...
  render.ErrorResponse:
    properties:
      error:
        example: 操作失败
        type: string
    type: object
  render.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  service.OpsJobCreateDTO:
    properties:
      description:
        example: 部署应用到生产环境
        type: string
      name:
        example: deploy-app
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Navy-NG 管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Navy-NG API
  version: "1.0"
paths:
...
  /fe-v1/ops-jobs:
    get:
      consumes:
      - application/json
      description: 获取运维任务列表，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页大小，默认10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取运维任务列表
      tags:
      - 运维任务
    post:
      consumes:
      - application/json
      description: 创建新的运维任务
      parameters:
      - description: 运维任务数据
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/service.OpsJobCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 创建运维任务
      tags:
      - 运维任务
  /fe-v1/ops-jobs/{id}:
    get:
      consumes:
      - application/json
      description: 根据任务ID获取运维任务的详细信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/render.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 获取运维任务详情
      tags:
      - 运维任务
  /fe-v1/ops-jobs/{id}/ws:
    get:
      consumes:
      - application/json
      description: 建立WebSocket连接以实时获取运维任务执行状态
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.ErrorResponse'
      summary: 处理WebSocket连接
      tags:
      - 运维任务
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"